requirements.txt
absl-py==0.4.1
astor==0.7.1
audioread==2.1.6
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
cffi==1.11.5
cycler==0.10.0
decorator==4.3.0
enum34==1.1.6
funcsigs==1.0.2
future==0.16.0
futures==3.2.0
gast==0.2.0
grpcio==1.14.1
h5py==2.8.0
joblib==0.12.2
kapre==0.1.3.1
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
kiwisolver==1.0.1
librosa==0.6.2
llvmlite==0.24.0
Markdown==2.6.11
matplotlib==2.2.3
mock==2.0.0
numba==0.39.0
numpy==1.14.5
pbr==4.2.0
protobuf==3.6.1
pycparser==2.18
pyparsing==2.2.0
pystoi==0.2.1
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
resampy==0.2.1
scikit-learn==0.19.2
scipy==1.1.0
singledispatch==3.4.0.3
six==1.11.0
SoundFile==0.10.2
subprocess32==3.5.2
tensorboard==1.10.0
tensorflow-gpu==1.10.1
termcolor==1.1.0
tqdm==4.25.0
Werkzeug==0.14.1
